fix: repair dashboard report widgets
